requirements.txt
sounddevice
numpy
scipy
pynput
pyperclip
python-dotenv
openai
PyQt6